+++ air/src/proof/mod.rs
@@ -6,11 +6,11 @@
 //! Contains STARK proof struct and associated components.
 
 use alloc::vec::Vec;
-use core::cmp;
 
 use crypto::{Hasher, MerkleTree};
 use fri::FriProof;
 use math::FieldElement;
+use security::{ConjecturedSecurity, ProvenSecurity};
 use utils::{ByteReader, Deserializable, DeserializationError, Serializable, SliceReader};
 
 use crate::{ProofOptions, TraceInfo};
@@ -27,17 +27,13 @@ pub use queries::Queries;
 mod ood_frame;
 pub use ood_frame::{OodFrame, TraceOodFrame};
 
+mod security;
+
 mod table;
 pub use table::Table;
 
 #[cfg(test)]
 mod tests;
-
-// CONSTANTS
-// ================================================================================================
-
-const GRINDING_CONTRIBUTION_FLOOR: u32 = 80;
-const MAX_PROXIMITY_PARAMETER: u64 = 1000;
 
 // PROOF
 // ================================================================================================
@@ -103,8 +99,8 @@ impl Proof {
     ///
     /// This is the conjecture on the security of the Toy problem (Conjecture 1)
     /// in https://eprint.iacr.org/2021/582.
-    pub fn security_level_conjectured<H: Hasher>(&self) -> ConjecturedSecurityBits {
-        get_conjectured_security(
+    pub fn security_level_conjectured<H: Hasher>(&self) -> ConjecturedSecurity {
+        ConjecturedSecurity::compute(
             self.context.options(),
             self.context.num_modulus_bits(),
             self.trace_info().length(),
@@ -115,8 +111,8 @@ impl Proof {
     ///
     /// Usually, the number of queries needed for provable security is 2x - 3x higher than
     /// the number of queries needed for conjectured security at the same security level.
-    pub fn security_level_proven<H: Hasher>(&self) -> ProvenSecurityBits {
-        get_proven_security(
+    pub fn security_level_proven<H: Hasher>(&self) -> ProvenSecurity {
+        ProvenSecurity::compute(
             self.context.options(),
             self.context.num_modulus_bits(),
             self.trace_info().length(),
@@ -207,717 +203,5 @@ impl Deserializable for Proof {
             gkr_proof: Option::<Vec<u8>>::read_from(source)?,
         };
         Ok(proof)
-    }
-}
-
-// HELPER FUNCTIONS
-// ================================================================================================
-
-pub struct ConjecturedSecurityBits(pub u32);
-
-pub struct ProvenSecurityBits {
-    unique_decoding: u32,
-    list_decoding: u32,
-}
-
-impl ProvenSecurityBits {
-    pub fn list_decoding(&self) -> u32 {
-        self.list_decoding
-    }
-
-    pub fn unique_decoding(&self) -> u32 {
-        self.unique_decoding
-    }
-}
-
-/// Computes conjectured security level for the specified proof parameters.
-fn get_conjectured_security(
-    options: &ProofOptions,
-    base_field_bits: u32,
-    trace_domain_size: usize,
-    collision_resistance: u32,
-) -> ConjecturedSecurityBits {
-    // compute max security we can get for a given field size
-    let field_size = base_field_bits * options.field_extension().degree();
-    let field_security = field_size - (trace_domain_size * options.blowup_factor()).ilog2();
-
-    // compute security we get by executing multiple query rounds
-    let security_per_query = options.blowup_factor().ilog2();
-    let mut query_security = security_per_query * options.num_queries() as u32;
-
-    // include grinding factor contributions only for proofs adequate security
-    if query_security >= GRINDING_CONTRIBUTION_FLOOR {
-        query_security += options.grinding_factor();
-    }
-
-    ConjecturedSecurityBits(cmp::min(
-        cmp::min(field_security, query_security) - 1,
-        collision_resistance,
-    ))
-}
-
-/// Estimates proven security level for the specified proof parameters.
-fn get_proven_security(
-    options: &ProofOptions,
-    base_field_bits: u32,
-    trace_domain_size: usize,
-    collision_resistance: u32,
-) -> ProvenSecurityBits {
-    let unique_decoding = cmp::min(
-        proven_security_protocol_unique_decoding(options, base_field_bits, trace_domain_size),
-        collision_resistance as u64,
-    ) as u32;
-
-    // determine the interval to which the which the optimal `m` belongs
-    let m_min: usize = 3;
-    let m_max = compute_upper_m(trace_domain_size);
-
-    // search for optimal `m` i.e., the one at which we maximize the number of security bits
-    let m_optimal = (m_min as u32..m_max as u32)
-        .max_by_key(|&a| {
-            proven_security_protocol_for_m(
-                options,
-                base_field_bits,
-                trace_domain_size,
-                a as usize,
-            )
-        })
-        .expect(
-            "Should not fail since m_max is larger than m_min for all trace sizes of length greater than 4",
-        );
-
-    let list_decoding = cmp::min(
-        proven_security_protocol_for_m(
-            options,
-            base_field_bits,
-            trace_domain_size,
-            m_optimal as usize,
-        ),
-        collision_resistance as u64,
-    ) as u32;
-
-    ProvenSecurityBits { unique_decoding, list_decoding }
-}
-
-/// Computes proven security level for the specified proof parameters for a fixed
-/// value of the proximity parameter m in the list-decoding regime.
-fn proven_security_protocol_for_m(
-    options: &ProofOptions,
-    base_field_bits: u32,
-    trace_domain_size: usize,
-    m: usize,
-) -> u64 {
-    let extension_field_bits = (base_field_bits * options.field_extension().degree()) as f64;
-    let num_fri_queries = options.num_queries() as f64;
-    let m = m as f64;
-    let rho = 1.0 / options.blowup_factor() as f64;
-    let alpha = (1.0 + 0.5 / m) * sqrt(rho);
-    // we use the blowup factor in order to bound the max degree
-    let max_deg = options.blowup_factor() as f64 + 1.0;
-    let lde_domain_size = (trace_domain_size * options.blowup_factor()) as f64;
-    let trace_domain_size = trace_domain_size as f64;
-    let num_openings = 2.0;
-
-    // we apply Theorem 2 in https://eprint.iacr.org/2024/1553, which is based on Theorem 8 in
-    // https://eprint.iacr.org/2022/1216.pdf and Theorem 5 in https://eprint.iacr.org/2021/582
-    // Note that the range of m needs to be restricted in order to ensure that eta, the slackness
-    // factor to the distance bound, is greater than 0.
-    // Determining the range of m is the responsibility of the calling function.
-    let mut epsilons_bits_neg = vec![];
-
-    // List size
-    let l = m / (rho - (2.0 * m / lde_domain_size));
-
-    // ALI related soundness error. Note that C here is equal to 1 because of our use of
-    // linear batching
-    let epsilon_1_bits_neg = -log2(l) + extension_field_bits;
-    epsilons_bits_neg.push(epsilon_1_bits_neg);
-
-    // DEEP related soundness error. Note that this uses that the denominator |F| - |D ∪ H|
-    // can be approximated by |F| for all practical domain sizes. We also use the blow-up factor
-    // as an upper bound for the maximal constraint degree.
-    let epsilon_2_bits_neg = -log2(
-        l * l * (max_deg * (trace_domain_size + num_openings - 1.0) + (trace_domain_size - 1.0)),
-    ) + extension_field_bits;
-    epsilons_bits_neg.push(epsilon_2_bits_neg);
-
-    // Computes FRI commit-phase (i.e., pre-query) soundness error.
-    // This considers only the first term given in eq. 7 in https://eprint.iacr.org/2022/1216.pdf,
-    // i.e. 0.5 * (m + 0.5)^7 * n^2 / (rho^1.5.q) as all other terms are negligible in comparison.
-    let epsilon_3_bits_neg = extension_field_bits
-        - log2((0.5 * powf(m + 0.5, 7.0) / powf(rho, 1.5)) * powf(lde_domain_size, 2.0));
-    epsilons_bits_neg.push(epsilon_3_bits_neg);
-
-    // epsilon_i for i in [3..(k-1)], where k is number of rounds, are also negligible
-
-    // Compute FRI query-phase soundness error
-    let epsilon_k_bits_neg = options.grinding_factor() as f64 - log2(powf(alpha, num_fri_queries));
-    epsilons_bits_neg.push(epsilon_k_bits_neg);
-
-    // return the round-by-round (RbR) soundness error
-    epsilons_bits_neg.into_iter().fold(f64::INFINITY, |a, b| a.min(b)) as u64
-}
-
-/// Computes proven security level for the specified proof parameters in the unique-decoding regime.
-fn proven_security_protocol_unique_decoding(
-    options: &ProofOptions,
-    base_field_bits: u32,
-    trace_domain_size: usize,
-) -> u64 {
-    let extension_field_bits = (base_field_bits * options.field_extension().degree()) as f64;
-    let num_fri_queries = options.num_queries() as f64;
-    let lde_domain_size = (trace_domain_size * options.blowup_factor()) as f64;
-    let trace_domain_size = trace_domain_size as f64;
-    let num_openings = 2.0;
-    let rho_plus = (trace_domain_size + num_openings) / lde_domain_size;
-    let alpha = (1.0 + rho_plus) * 0.5;
-    // we use the blowup factor in order to bound the max degree
-    let max_deg = options.blowup_factor() as f64 + 1.0;
-
-    // we apply Theorem 3 in https://eprint.iacr.org/2024/1553
-    let mut epsilons_bits_neg = vec![];
-
-    // ALI related soundness error
-    let epsilon_1_bits_neg = extension_field_bits;
-    epsilons_bits_neg.push(epsilon_1_bits_neg);
-
-    // DEEP related soundness error. Note that this uses that the denominator |F| - |D ∪ H|
-    // can be approximated by |F| for all practical domain sizes. We also use the blow-up factor
-    // as an upper bound for the maximal constraint degree
-    let epsilon_2_bits_neg =
-        -log2(max_deg * (trace_domain_size + num_openings - 1.0) + (trace_domain_size - 1.0))
-            + extension_field_bits;
-    epsilons_bits_neg.push(epsilon_2_bits_neg);
-
-    // Computes FRI commit-phase (i.e., pre-query) soundness error
-    let epsilon_3_bits_neg = extension_field_bits - log2(lde_domain_size);
-    epsilons_bits_neg.push(epsilon_3_bits_neg);
-
-    // epsilon_i for i in [3..(k-1)], where k is number of rounds
-    let folding_factor = options.to_fri_options().folding_factor() as f64;
-    let num_fri_layers = options.to_fri_options().num_fri_layers(lde_domain_size as usize);
-    let epsilon_i_min_bits_neg = (0..num_fri_layers)
-        .map(|_| extension_field_bits - log2((folding_factor - 1.0) * (lde_domain_size + 1.0)))
-        .fold(f64::INFINITY, |a, b| a.min(b));
-    epsilons_bits_neg.push(epsilon_i_min_bits_neg);
-
-    // Compute FRI query-phase soundness error
-    let epsilon_k_bits_neg = options.grinding_factor() as f64 - log2(powf(alpha, num_fri_queries));
-    epsilons_bits_neg.push(epsilon_k_bits_neg);
-
-    // return the round-by-round (RbR) soundness error
-    epsilons_bits_neg.into_iter().fold(f64::INFINITY, |a, b| a.min(b)) as u64
-}
-
-// HELPER FUNCTIONS
-// ================================================================================================
-
-/// Computes the largest proximity parameter m such that eta is greater than 0 in the proof of
-/// Theorem 1 in https://eprint.iacr.org/2021/582. See Theorem 2 in https://eprint.iacr.org/2024/1553
-/// and its proof for more on this point.
-///
-/// The bound on m in Theorem 2 in https://eprint.iacr.org/2024/1553 is sufficient but we can use
-/// the following to compute a better bound.
-fn compute_upper_m(h: usize) -> f64 {
-    let h = h as f64;
-    let ratio = (h + 2.0) / h;
-    let m_max = ceil(1.0 / (2.0 * (sqrt(ratio) - 1.0)));
-    assert!(m_max >= h / 2.0, "the bound in the theorem should be tighter");
-
-    // We cap the range to 1000 as the optimal m value will be in the lower range of [m_min, m_max]
-    // since increasing m too much will lead to a deterioration in the FRI commit soundness making
-    // any benefit gained in the FRI query soundess mute.
-    cmp::min(m_max as u64, MAX_PROXIMITY_PARAMETER) as f64
-}
-
-#[cfg(feature = "std")]
-pub fn log2(value: f64) -> f64 {
-    value.log2()
-}
-
-#[cfg(not(feature = "std"))]
-pub fn log2(value: f64) -> f64 {
-    libm::log2(value)
-}
-
-#[cfg(feature = "std")]
-pub fn sqrt(value: f64) -> f64 {
-    value.sqrt()
-}
-
-#[cfg(not(feature = "std"))]
-pub fn sqrt(value: f64) -> f64 {
-    libm::sqrt(value)
-}
-
-#[cfg(feature = "std")]
-pub fn powf(value: f64, exp: f64) -> f64 {
-    value.powf(exp)
-}
-
-#[cfg(not(feature = "std"))]
-pub fn powf(value: f64, exp: f64) -> f64 {
-    libm::pow(value, exp)
-}
-
-#[cfg(feature = "std")]
-pub fn ceil(value: f64) -> f64 {
-    value.ceil()
-}
-
-#[cfg(not(feature = "std"))]
-pub fn ceil(value: f64) -> f64 {
-    libm::ceil(value)
-}
-
-#[cfg(test)]
-mod prove_security_tests {
-    use math::{fields::f64::BaseElement, StarkField};
-
-    use super::ProofOptions;
-    use crate::{
-        proof::{get_proven_security, ProvenSecurityBits},
-        FieldExtension,
-    };
-
-    #[test]
-    fn get_100_bits_security() {
-        let field_extension = FieldExtension::Quadratic;
-        let base_field_bits = BaseElement::MODULUS_BITS;
-        let fri_folding_factor = 2;
-        let fri_remainder_max_degree = 127;
-        let grinding_factor = 20;
-        let blowup_factor = 4;
-        let num_queries = 119;
-        let collision_resistance = 128;
-        let trace_length = 2_usize.pow(20);
-
-        let mut options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding, list_decoding } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(unique_decoding, 100);
-        assert_eq!(list_decoding, 69);
-
-        // increasing the queries does not help the LDR case
-        let num_queries = 150;
-
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding: _, list_decoding } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(list_decoding, 69);
-
-        // increasing the extension degree does help and we then need fewer queries by virtue
-        // of being in LDR
-        let field_extension = FieldExtension::Cubic;
-        let num_queries = 81;
-
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding: _, list_decoding } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(list_decoding, 100);
-    }
-
-    #[test]
-    fn unique_decoding_folding_factor_effect() {
-        let field_extension = FieldExtension::Quadratic;
-        let base_field_bits = BaseElement::MODULUS_BITS;
-        let fri_folding_factor = 2;
-        let fri_remainder_max_degree = 7;
-        let grinding_factor = 16;
-        let blowup_factor = 8;
-        let num_queries = 123;
-        let collision_resistance = 128;
-        let trace_length = 2_usize.pow(8);
-
-        let mut options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding, list_decoding: _ } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(unique_decoding, 116);
-
-        let fri_folding_factor = 4;
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding, list_decoding: _ } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(unique_decoding, 115);
-    }
-
-    #[test]
-    fn unique_versus_list_decoding_rate_effect() {
-        let field_extension = FieldExtension::Quadratic;
-        let base_field_bits = BaseElement::MODULUS_BITS;
-        let fri_folding_factor = 2;
-        let fri_remainder_max_degree = 7;
-        let grinding_factor = 20;
-        let blowup_factor = 2;
-        let num_queries = 195;
-        let collision_resistance = 128;
-        let trace_length = 2_usize.pow(8);
-
-        let mut options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding, list_decoding: _ } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(unique_decoding, 100);
-
-        // when the rate is large, going to a larger extension field in order to make full use of
-        // being in the LDR might not always be justified
-
-        // we increase the extension degree
-        let field_extension = FieldExtension::Cubic;
-        // and we reduce the number of required queries to reach the target level, but this is
-        // a relatively small, approximately 16%, reduction
-        let num_queries = 163;
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding: _, list_decoding } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(list_decoding, 100);
-
-        // reducing the rate further changes things
-        let field_extension = FieldExtension::Quadratic;
-        let blowup_factor = 4;
-        let num_queries = 119;
-
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding, list_decoding: _ } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(unique_decoding, 100);
-
-        // the improvement is now at approximately 32%
-        let field_extension = FieldExtension::Cubic;
-        let num_queries = 81;
-
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding: _, list_decoding } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(list_decoding, 100);
-    }
-
-    #[test]
-    fn get_96_bits_security() {
-        let field_extension = FieldExtension::Cubic;
-        let base_field_bits = BaseElement::MODULUS_BITS;
-        let fri_folding_factor = 8;
-        let fri_remainder_max_degree = 127;
-        let grinding_factor = 20;
-        let blowup_factor = 4;
-        let num_queries = 80;
-        let collision_resistance = 128;
-        let trace_length = 2_usize.pow(18);
-
-        let mut options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding: _, list_decoding } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(list_decoding, 99);
-
-        // increasing the blowup factor should increase the bits of security gained per query
-        let blowup_factor = 8;
-        let num_queries = 53;
-
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding: _, list_decoding } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(list_decoding, 99);
-    }
-
-    #[test]
-    fn get_128_bits_security() {
-        let field_extension = FieldExtension::Cubic;
-        let base_field_bits = BaseElement::MODULUS_BITS;
-        let fri_folding_factor = 8;
-        let fri_remainder_max_degree = 127;
-        let grinding_factor = 20;
-        let blowup_factor = 8;
-        let num_queries = 85;
-        let collision_resistance = 128;
-        let trace_length = 2_usize.pow(18);
-
-        let mut options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding: _, list_decoding } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(list_decoding, 128);
-
-        // increasing the blowup factor should increase the bits of security gained per query
-        let blowup_factor = 16;
-        let num_queries = 65;
-
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding: _, list_decoding } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(list_decoding, 128);
-    }
-
-    #[test]
-    fn extension_degree() {
-        let field_extension = FieldExtension::Quadratic;
-        let base_field_bits = BaseElement::MODULUS_BITS;
-        let fri_folding_factor = 8;
-        let fri_remainder_max_degree = 127;
-        let grinding_factor = 20;
-        let blowup_factor = 8;
-        let num_queries = 85;
-        let collision_resistance = 128;
-        let trace_length = 2_usize.pow(18);
-
-        let mut options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding: _, list_decoding } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(list_decoding, 69);
-
-        // increasing the extension degree improves the FRI commit phase soundness error and permits
-        // reaching 128 bits security
-        let field_extension = FieldExtension::Cubic;
-
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits { unique_decoding: _, list_decoding } =
-            get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert_eq!(list_decoding, 128);
-    }
-
-    #[test]
-    fn trace_length() {
-        let field_extension = FieldExtension::Cubic;
-        let base_field_bits = BaseElement::MODULUS_BITS;
-        let fri_folding_factor = 8;
-        let fri_remainder_max_degree = 127;
-        let grinding_factor = 20;
-        let blowup_factor = 8;
-        let num_queries = 80;
-        let collision_resistance = 128;
-        let trace_length = 2_usize.pow(20);
-
-        let mut options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits {
-            unique_decoding: _,
-            list_decoding: security_1,
-        } = get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        let trace_length = 2_usize.pow(16);
-
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits {
-            unique_decoding: _,
-            list_decoding: security_2,
-        } = get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert!(security_1 < security_2);
-    }
-
-    #[test]
-    fn num_fri_queries() {
-        let field_extension = FieldExtension::Cubic;
-        let base_field_bits = BaseElement::MODULUS_BITS;
-        let fri_folding_factor = 8;
-        let fri_remainder_max_degree = 127;
-        let grinding_factor = 20;
-        let blowup_factor = 8;
-        let num_queries = 60;
-        let collision_resistance = 128;
-        let trace_length = 2_usize.pow(20);
-
-        let mut options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits {
-            unique_decoding: _,
-            list_decoding: security_1,
-        } = get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        let num_queries = 80;
-
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits {
-            unique_decoding: _,
-            list_decoding: security_2,
-        } = get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert!(security_1 < security_2);
-    }
-
-    #[test]
-    fn blowup_factor() {
-        let field_extension = FieldExtension::Cubic;
-        let base_field_bits = BaseElement::MODULUS_BITS;
-        let fri_folding_factor = 8;
-        let fri_remainder_max_degree = 127;
-        let grinding_factor = 20;
-        let blowup_factor = 8;
-        let num_queries = 30;
-        let collision_resistance = 128;
-        let trace_length = 2_usize.pow(20);
-
-        let mut options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits {
-            unique_decoding: _,
-            list_decoding: security_1,
-        } = get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        let blowup_factor = 16;
-
-        options = ProofOptions::new(
-            num_queries,
-            blowup_factor,
-            grinding_factor,
-            field_extension,
-            fri_folding_factor as usize,
-            fri_remainder_max_degree as usize,
-        );
-        let ProvenSecurityBits {
-            unique_decoding: _,
-            list_decoding: security_2,
-        } = get_proven_security(&options, base_field_bits, trace_length, collision_resistance);
-
-        assert!(security_1 < security_2);
     }
 }
