+++ verifier/src/lib.rs
@@ -34,12 +34,12 @@ extern crate alloc;
 use alloc::{string::ToString, vec::Vec};
 use core::cmp;
 
-use air::{proof::ConjecturedSecurityBits, AuxRandElements, GkrVerifier};
 pub use air::{
     proof::Proof, Air, AirContext, Assertion, BoundaryConstraint, BoundaryConstraintGroup,
     ConstraintCompositionCoefficients, ConstraintDivisor, DeepCompositionCoefficients,
     EvaluationFrame, FieldExtension, ProofOptions, TraceInfo, TransitionConstraintDegree,
 };
+use air::{AuxRandElements, GkrVerifier};
 pub use crypto;
 use crypto::{ElementHasher, Hasher, RandomCoin, VectorCommitment};
 use fri::FriVerifier;
@@ -360,26 +360,20 @@ impl AcceptableOptions {
     pub fn validate<H: Hasher>(&self, proof: &Proof) -> Result<(), VerifierError> {
         match self {
             AcceptableOptions::MinConjecturedSecurity(minimal_security) => {
-                let ConjecturedSecurityBits(proof_security) =
-                    proof.security_level_conjectured::<H>();
-                if proof_security < *minimal_security {
+                let conjectured_security = proof.security_level_conjectured::<H>();
+                if !conjectured_security.is_at_least(*minimal_security) {
                     return Err(VerifierError::InsufficientConjecturedSecurity(
                         *minimal_security,
-                        proof_security,
+                        conjectured_security.bits(),
                     ));
                 }
             },
             AcceptableOptions::MinProvenSecurity(minimal_security) => {
                 let proven_security = proof.security_level_proven::<H>();
-                if proven_security.list_decoding() < *minimal_security
-                    && proven_security.unique_decoding() < *minimal_security
-                {
+                if !proven_security.is_at_least(*minimal_security) {
                     return Err(VerifierError::InsufficientProvenSecurity(
                         *minimal_security,
-                        cmp::max(
-                            proven_security.list_decoding(),
-                            proven_security.unique_decoding(),
-                        ),
+                        cmp::max(proven_security.ldr_bits(), proven_security.udr_bits()),
                     ));
                 }
             },
